tests/services/test_review_service.py
"""
test_review_service.py

Unit tests for the Review Service.

This module verifies the correct behavior of the ReviewService, focusing on:

- Creating reviews for menu items or restaurants
- Preventing duplicate reviews from the same user
- Retrieving reviews by item, restaurant, or user
- Updating and deleting reviews
- Calculating average ratings

Mocks are used to simulate database operations and isolate the service logic.
"""

import pytest
from pytest_mock import MockerFixture

@pytest.fixture
def mock_db():
    pass

@pytest.fixture
def service(mock_db):
    pass

def test_create_review(service):
    """
    Test creating a review with valid input.
    """
    pass

def test_prevent_duplicate_review(service):
    """
    Test that the system prevents users from submitting multiple reviews for the same item.
    """
    pass

def test_get_reviews_by_item(service):
    """
    Test retrieving all reviews for a specific menu item.
    """
    pass

def test_get_reviews_by_user(service):
    """
    Test retrieving all reviews submitted by a specific user.
    """
    pass

def test_update_review(service):
    """
    Test updating the content or rating of an existing review.
    """
    pass

def test_delete_review(service):
    """
    Test deleting a review by its ID.
    """
    pass

def test_calculate_average_rating(service):
    """
    Test calculating the average rating for a menu item or restaurant.
    """
    pass
